feat: language switcher adjustments
