+++ client/src/app/globals.scss
@@ -7,6 +7,7 @@
 }
 
 body {
+  color: #fff;
   background: linear-gradient(90deg,
       rgba(48, 16, 255, 1) 0%,
       rgba(100, 115, 255, 1) 100%);
@@ -33,4 +34,14 @@ h1 {
   padding-left: 15px;
   padding-right: 15px;
   max-width: 1240px;
+}
+
+.btn-reset {
+  border: none;
+  font-family: inherit;
+  padding: 0;
+  outline: none;
+  background-color: transparent;
+  cursor: pointer;
+  color: #fff;
 }
