+++ client/src/components/Layout/Header/Header.module.scss
@@ -7,3 +7,8 @@
   rgb(97, 76, 240) 0%,
   rgb(127, 139, 245) 100%);
 }
+
+.row {
+  display: flex;
+  gap: 20px;
+}
