+++ client/src/components/Layout/Header/Header.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 
 import { MainMenu } from './MainMenu';
 import { ProfileButton } from './ProfileButton';
+import { LanguageSwitcher } from './LanguageSwitcher';
 
 import styles from './Header.module.scss';
 
@@ -10,7 +11,10 @@ export const Header = () => {
   return (
     <header className={clsx(styles.header, 'container')}>
       <MainMenu />
-      <ProfileButton />
+      <div className={styles.row}>
+        <LanguageSwitcher />
+        <ProfileButton />
+      </div>
     </header>
   );
 };
