+++ client/src/components/TodoList/TodoList.tsx
@@ -7,7 +7,6 @@ import { Todo as TodoInterface } from '@/shared/types';
 
 import { TodoForm } from '../TodoForm';
 import { Todos } from '../Todos';
-import { LanguageSwitcher } from '../LanguageSwitcher';
 
 import styles from './TodoList.module.scss';
 
@@ -31,7 +30,6 @@ export const TodoList: FC<TodoListProps> = ({ dictionary }) => {
   return (
     <div className={styles.todoApp}>
       <h1>{dictionary.todo.todoListTitle}</h1>
-      <LanguageSwitcher />
       <TodoForm />
       <Todos todos={todos} />
     </div>
